Deleted __php__
These files are a crunch to avoid placing path to document root of the
server to every script manually. Instead, been placed in every directory
of the project down to the root, these  files automatically find root
and declare it as a constant.  I'll try to avoid these, but still...
placing hardcoded path into .php scripts is ugly, but settings.ini can't
help me, because it's still a file that needs to be loaded :)